Fix focus used for calibration.
The script stores 135 in the calibration output, but actually used 0.

+++ calibrate.py
@@ -642,7 +642,7 @@ class Main:
                         if self.args.rgbLensPosition:
                             cam_node.initialControl.setManualFocus(int(self.args.rgbLensPosition[stringToCam[cam_id].name.lower()]))
                         else:
-                            cam_node.initialControl.setManualFocusRaw(int(135 / 255))
+                            cam_node.initialControl.setManualFocusRaw(135 / 255)
 
                     self.control_queue = cam_node.inputControl.createInputQueue()
                     sensorName = cam_info['sensorName']
